requirements.txt
aiohttp==3.11.18
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses-json==0.6.7
faiss-cpu==1.11.0
fsspec==2025.3.2
huggingface-hub==0.31.1
idna==3.10
joblib==1.5.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-huggingface==0.2.0
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.42
marshmallow==3.26.1
numpy==2.2.5
ollama==0.4.8
orjson==3.10.18
packaging==24.2
pydantic==2.11.4
pydantic-settings==2.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
SQLAlchemy==2.0.40
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zstandard==0.23.0
